LayoutTests/fast/css/parsing-css-not-5-expected.txt

Test the parsing of :not(selectorList) for querySelector and style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


1 invalid selectors
PASS document.querySelector(":not(::first-letter)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(::first-line)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(::before)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(::after)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:first-letter)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:first-line)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:before)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:after)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not()") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(    )") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(,)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(.123class)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(#123id)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not([])") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(())") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not())") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not({})") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(})") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:not())") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:matches())") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:nth-child())") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:nth-child(2n of))") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:not()") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:matches()") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":not(:nth-child(2n+1 of)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS successfullyParsed is true

TEST COMPLETE
